+++ gear/form_qc_checker/src/python/form_qc_app/datastore.py
@@ -4,15 +4,19 @@ import logging
 from typing import Any, Dict, List, Optional
 
 from centers.nacc_group import NACCGroup
-from configs.ingest_configs import ModuleConfigs
+from configs.ingest_configs import FormProjectConfigs, ModuleConfigs
 from datastore.forms_store import FormsStore
 from flywheel_adaptor.flywheel_proxy import FlywheelProxy, ProjectAdaptor, ProjectError
-from keys.keys import DefaultValues
+from keys.keys import DefaultValues, MetadataKeys
 from nacc_common.field_names import FieldNames
 from nacc_form_validator.datastore import Datastore
 from rxnorm.rxnorm_connection import RxcuiStatus, RxNormConnection
 
 log = logging.getLogger(__name__)
+
+
+class DatastoreException(Exception):
+    pass
 
 
 class DatastoreHelper(Datastore):
@@ -26,35 +30,44 @@ class DatastoreHelper(Datastore):
         *,
         pk_field: str,
         adcid: int,
+        module: str,
         group_id: str,
         project: ProjectAdaptor,
         proxy: FlywheelProxy,
         admin_group: NACCGroup,
         module_configs: ModuleConfigs,
-        legacy_label: str,
+        form_project_configs: FormProjectConfigs,
     ):
         """
 
         Args:
             pk_field: primary key field to uniquely identify a participant
             adcid: Center's ADCID
+            module: module label
             group: Flywheel group id
             project: Flywheel project adaptor
             proxy: Flywheel proxy object
             admin_group: Flywheel admin group
             module_configs: form ingest configs for the module
-            legacy_label: legacy project label
+            form_project_configs: form ingest configs for all modules for the project
         """
 
         super().__init__(pk_field, module_configs.date_field)
 
         self.__proxy = proxy
         self.__adcid = adcid
+        self.__module = module.upper()
         self.__gid = group_id
         self.__project = project
         self.__admin_group = admin_group
         self.__module_configs = module_configs
+        self.__form_configs = form_project_configs
 
+        legacy_label = (
+            self.__form_configs.legacy_project_label
+            if self.__form_configs.legacy_project_label
+            else DefaultValues.LEGACY_PRJ_LABEL
+        )
         self.__forms_store = FormsStore(
             ingest_project=self.__project,
             legacy_project=self.__get_legacy_project(legacy_label),
@@ -63,7 +76,13 @@ class DatastoreHelper(Datastore):
         self.__current_adcids = self.__pull_adcids_list()
 
         # cache for grabbing previous records
-        self.__prev_visits: Dict[str, Any] = {}
+        self.__prev_visits: Dict[str, Any] = {}  # by subject, module, date
+
+        # cache for initial visits
+        self.__initial_visits: Dict[str, Any] = {}  # by subject and module
+
+        # cache for UDS IVP visits
+        self.__uds_ivp_visits: Dict[str, Any] = {}  # by subject
 
     def __pull_adcids_list(self) -> Optional[List[int]]:
         """Pull the list of ADCIDs from the admin group metadata project.
@@ -117,18 +136,16 @@ class DatastoreHelper(Datastore):
             )
             return None
 
-    def __get_previous_visits(
-        self, current_record: Dict[str, str]
-    ) -> Optional[List[Dict[str, str]]]:
-        """Retrieve the list of previous visits for the specified participant.
+    def __validate_current_record(self, current_record: Dict[str, Any]) -> bool:
+        """Validate the current record and ensure it has all required fields.
+        Technically this should not happen, but run as a sanity check.
 
         Args:
             current_record: record currently being validated
 
         Returns:
-            List[Dict[str, str]]: previous visit records if found, else None
+            bool: whether or not record has all required fields
         """
-
         required_fields = [self.pk_field, self.orderby, FieldNames.MODULE]
 
         found_all = True
@@ -137,14 +154,32 @@ class DatastoreHelper(Datastore):
                 log.error(
                     (
                         "Field %s not set in current visit data, "
-                        "cannot retrieve the previous visits"
+                        "cannot retrieve the previous/initial visits"
                     ),
                     field,
                 )
                 found_all = False
 
-        # this cannot happen, just a sanity check
-        if not found_all:
+        if found_all and self.__module != current_record[FieldNames.MODULE].upper():
+            raise DatastoreException(
+                f"Datastore module {self.__module} mismatch with the "
+                f"module specified in the record {current_record[FieldNames.MODULE]}"
+            )
+
+        return found_all
+
+    def __get_previous_visits(
+        self, current_record: Dict[str, Any]
+    ) -> Optional[List[Dict[str, Any]]]:
+        """Retrieve the list of previous visits for the specified participant.
+
+        Args:
+            current_record: record currently being validated
+
+        Returns:
+            List[Dict[str, Any]]: previous visit records if found, else None
+        """
+        if not self.__validate_current_record(current_record):
             return None
 
         subject_lbl = current_record[self.pk_field]
@@ -188,13 +223,13 @@ class DatastoreHelper(Datastore):
 
         # if no previous visits found in the current project, check the legacy project
         legacy_module = (
-            self.__module_configs.legacy_module
+            self.__module_configs.legacy_module.label
             if self.__module_configs.legacy_module
             else module
         )
         legacy_date = (
-            self.__module_configs.legacy_date
-            if self.__module_configs.legacy_date
+            self.__module_configs.legacy_module.date_field
+            if self.__module_configs.legacy_module
             else self.orderby
         )
 
@@ -224,9 +259,124 @@ class DatastoreHelper(Datastore):
 
         return legacy_visits
 
+    def __get_initial_visit(
+        self, current_record: Dict[str, Any]
+    ) -> Optional[Dict[str, Any]]:
+        """Retrieve the initial visit for the specified participant. Return the
+        IVP packet for the modules that has only one initial packet, else
+        return the first record sorted by visit date or form date.
+
+        Args:
+            current_record: record currently being validated
+
+        Returns:
+            Dict[str, Any]: initial visit record if found, else None
+        """
+
+        subject_lbl = current_record[self.pk_field]
+        module = current_record[FieldNames.MODULE].upper()
+        ivp_codes = self.__module_configs.initial_packets
+        date_field = self.__module_configs.date_field
+
+        initial_visits = self.__forms_store.query_form_data(
+            subject_lbl=subject_lbl,
+            module=module,
+            legacy=False,
+            search_col=FieldNames.PACKET,
+            search_val=ivp_codes,
+            search_op=DefaultValues.FW_SEARCH_OR,
+            qc_gear=DefaultValues.QC_GEAR,
+            extra_columns=[date_field],
+        )
+
+        if not initial_visits:
+            if self.__module_configs.legacy_module:
+                date_field = self.__module_configs.legacy_module.date_field
+                if self.__module_configs.legacy_module.initial_packets:
+                    ivp_codes = self.__module_configs.legacy_module.initial_packets
+
+            initial_visits = self.__forms_store.query_form_data(
+                subject_lbl=subject_lbl,
+                module=module,
+                legacy=True,
+                search_col=FieldNames.PACKET,
+                search_val=ivp_codes,
+                search_op=DefaultValues.FW_SEARCH_OR,
+                qc_gear=DefaultValues.QC_GEAR,
+                extra_columns=[date_field],
+            )
+
+        if not initial_visits:
+            log.warning("No initial visit found for %s, module %s", subject_lbl, module)
+            return None
+
+        if len(initial_visits) > 1:
+            log.warning(
+                "Multiple initial visits found for %s, module %s", subject_lbl, module
+            )
+            date_col_key = f"{MetadataKeys.FORM_METADATA_PATH}.{date_field}"
+            initial_visits = sorted(initial_visits, key=lambda d: d[date_col_key])
+
+        return initial_visits[0]
+
+    def __get_uds_ivp_visit(
+        self, current_record: Dict[str, Any], uds_configs: ModuleConfigs
+    ) -> Optional[Dict[str, Any]]:
+        """Retrieve the UDS IVP for the specified participant if present.
+
+        Args:
+            current_record: record currently being validated
+            uds_configs: form ingest configs for UDS module
+
+        Returns:
+            Dict[str, Any]: UDS IVP record if found, else None
+        """
+
+        subject_lbl = current_record[self.pk_field]
+        ivp_codes = uds_configs.initial_packets
+        date_field = uds_configs.date_field
+
+        uds_ivp_visits = self.__forms_store.query_form_data(
+            subject_lbl=subject_lbl,
+            module=DefaultValues.UDS_MODULE,
+            legacy=False,
+            search_col=FieldNames.PACKET,
+            search_val=ivp_codes,
+            search_op=DefaultValues.FW_SEARCH_OR,
+            qc_gear=DefaultValues.QC_GEAR,
+            extra_columns=[date_field],
+        )
+
+        if not uds_ivp_visits:
+            if uds_configs.legacy_module:
+                date_field = uds_configs.legacy_module.date_field
+                if uds_configs.legacy_module.initial_packets:
+                    ivp_codes = uds_configs.legacy_module.initial_packets
+
+            uds_ivp_visits = self.__forms_store.query_form_data(
+                subject_lbl=subject_lbl,
+                module=DefaultValues.UDS_MODULE,
+                legacy=True,
+                search_col=FieldNames.PACKET,
+                search_val=ivp_codes,
+                search_op=DefaultValues.FW_SEARCH_OR,
+                qc_gear=DefaultValues.QC_GEAR,
+                extra_columns=[date_field],
+            )
+
+        if not uds_ivp_visits:
+            log.warning("No UDS IVP record found for %s", subject_lbl)
+            return None
+
+        if len(uds_ivp_visits) > 1:
+            log.error("Multiple UDS IVP records found for %s", subject_lbl)
+            return None
+
+        return uds_ivp_visits[0]
+
     def get_previous_record(
-        self, current_record: Dict[str, str]
-    ) -> Optional[Dict[str, str]]:
+        self, current_record: Dict[str, Any]
+    ) -> Optional[Dict[str, Any]]:
         """Overriding the abstract method, get the previous visit record for
         the specified participant.
 
@@ -247,20 +397,38 @@ class DatastoreHelper(Datastore):
             acq_id=latest_rec_info["file.parents.acquisition"],
         )
 
+    def __check_nonempty(
+        self,
+        ignore_empty_fields: Optional[List[str]] = None,
+        visit_data: Optional[Dict[str, Any]] = None,
+    ) -> bool:
+        """Returns whether all specified fields are not empty in visit_data.
+
+        Args:
+            ignore_empty_fields: Field(s) to check for blanks
+            visit_data: Record to check
+        """
+        if not ignore_empty_fields:
+            return True
+
+        if not visit_data:
+            return False
+
+        return all(visit_data.get(field) for field in ignore_empty_fields)
+
     def get_previous_nonempty_record(
-        self, current_record: Dict[str, str], fields: List[str]
-    ) -> Optional[Dict[str, str]]:
+        self, current_record: Dict[str, Any], ignore_empty_fields: List[str]
+    ) -> Optional[Dict[str, Any]]:
         """Overriding the abstract method to return the previous record where
         all fields are NOT empty for the specified participant.
 
         Args:
             current_record: Record currently being validated
-            fields: Field(s) to check for blanks
+            ignore_empty_fields: Field(s) to check for blanks
 
         Returns:
             Dict[str, str]: Previous non-empty record if found, else None
         """
-
         prev_visits = self.__get_previous_visits(current_record)
         if not prev_visits:
             return None
@@ -270,20 +438,103 @@ class DatastoreHelper(Datastore):
                 file_name=visit["file.name"], acq_id=visit["file.parents.acquisition"]
             )
 
-            if not visit_data:
-                continue
-
-            found_all = True
-            for field in fields:
-                if not visit_data.get(field):
-                    found_all = False
-                    break
-
-            if found_all:
+            if self.__check_nonempty(ignore_empty_fields, visit_data):
                 return visit_data
 
-        log.warning("No previous visit found with non-empty values for %s", fields)
+        log.warning(
+            "No previous visit found with non-empty values for %s", ignore_empty_fields
+        )
         return None
+
+    def get_initial_record(
+        self, current_record: Dict[str, Any]
+    ) -> Optional[Dict[str, Any]]:
+        """Overriding the abstract method, get the initial visit record for the
+        specified participant.
+
+        Args:
+            current_record: record currently being validated
+
+        Returns:
+            Dict[str, Any]: initial visit record if found, None otherwise
+        """
+
+        if not self.__validate_current_record(current_record):
+            return None
+
+        subject_lbl = current_record[self.pk_field]
+        module = current_record[FieldNames.MODULE].upper()
+
+        # see if we've already cached the initial record
+        cached_initial_record = self.__initial_visits.get(subject_lbl, {}).get(module)
+
+        # will be None if we've never looked for it, and empty dict if
+        # we tried looking for it but it could not be found
+        if cached_initial_record is None:
+            initial_record = self.__get_initial_visit(current_record)
+
+            if not initial_record:
+                initial_record = {}
+
+            self.__initial_visits.update({subject_lbl: {module: initial_record}})
+        else:
+            log.info("Already searched for initial visit, using cached records")
+            initial_record = cached_initial_record
+
+        if not initial_record:
+            return None
+
+        return self.__forms_store.get_visit_data(
+            file_name=initial_record["file.name"],
+            acq_id=initial_record["file.parents.acquisition"],
+        )
+
+    def get_uds_ivp_record(
+        self, current_record: Dict[str, Any]
+    ) -> Optional[Dict[str, Any]]:
+        """Overriding the abstract method, get the UDS IVP record for the
+        specified participant.
+
+        Args:
+            current_record: record currently being validated
+
+        Returns:
+            Dict[str, Any]: UDS IVP record if found, else None
+        """
+
+        if not self.__validate_current_record(current_record):
+            return None
+
+        uds_configs = self.__form_configs.module_configs.get(DefaultValues.UDS_MODULE)
+
+        if not uds_configs:
+            raise DatastoreException("Module configurations not found for UDS module")
+
+        subject_lbl = current_record[self.pk_field]
+
+        # see if we've already cached the UDS IVP record
+        cached_uds_ivp = self.__uds_ivp_visits.get(subject_lbl)
+
+        # will be None if we've never looked for it, and empty dict if
+        # we tried looking for it but it could not be found
+        if cached_uds_ivp is None:
+            uds_ivp_record = self.__get_uds_ivp_visit(current_record, uds_configs)
+
+            if not uds_ivp_record:
+                uds_ivp_record = {}
+
+            self.__uds_ivp_visits.update({subject_lbl: uds_ivp_record})
+        else:
+            log.info("Already searched for UDS IVP visit, using cached records")
+            uds_ivp_record = cached_uds_ivp
+
+        if not uds_ivp_record:
+            return None
+
+        return self.__forms_store.get_visit_data(
+            file_name=uds_ivp_record["file.name"],
+            acq_id=uds_ivp_record["file.parents.acquisition"],
+        )
 
     def is_valid_rxcui(self, drugid: int) -> bool:
         """Overriding the abstract method, check whether a given drug ID is