+++ docs/form_qc_checker/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 All notable changes to this gear are documented in this file.
+
+## 1.6.0
+* Updates `nacc-form-validator` to `0.6.0`
+  * Implements the new `get_initial_record` and `get_uds_ivp_record` Datastore methods
+* Rebuilt for form ingest configs update
   
 ## 1.5.7
 * Supports having study suffix in retrospective-form project label