+++ docs/form_transformer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.6.0
+* Adds NP vs UDS checks
+* Rebuilt for form ingest configs update
+  
 ## 1.5.1
 * Supports having study suffix in retrospective-form project label
 * Rebuilt for reset error log content in read failure